fixing a bunch of stuff in monitor.go
There was a problem with how I was originally trying to start the monitor. It
was do to the go-client caches not being ready and trying to use them too soon.
The solution was to implement manager.Runnable and let the controller-runtime
manager start the monitor, which is done after client caches have started.

I also made several changes to 1) detect if there was an out of band change and
2) trigger reconciliation to put the Reaper object back in the desired state.